Modified track_car.cpp to compile against OpenCV on my desktop computer, versus my laptop.  It turns out that the version that comes with 12.04, by default, is 2.3, while the version that comes with ROS Hydro is 2.4.9.  Version 2.3 is on my laptop, and 2.4.9 is on my desktop.
I have also modified track_car.cpp and sift.py to have identical print
statements.  This allowed me to find that I was truncating the center
coordinates for the training image to integers, in sift.py, which I didn't have
to do.  The two versions now print nearly identical numbers.  (The only
differences appear to be due to rounding.)